+++ src/main/java/fuzs/forgeconfigapiport/impl/integration/configured/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package fuzs.forgeconfigapiport.impl.integration.configured;
+
+import org.jetbrains.annotations.ApiStatus;