+++ src/main/java/fuzs/forgeconfigapiport/impl/network/client/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package fuzs.forgeconfigapiport.impl.network.client;
+
+import org.jetbrains.annotations.ApiStatus;