+++ src/main/java/fuzs/forgeconfigapiport/impl/util/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package fuzs.forgeconfigapiport.impl.util;
+
+import org.jetbrains.annotations.ApiStatus;