+++ src/main/java/net/minecraftforge/client/gui/widget/package-info.java
@@ -1,4 +1,0 @@
-@ApiStatus.Experimental
-package net.minecraftforge.client.gui.widget;
-
-import org.jetbrains.annotations.ApiStatus;