+++ src/main/java/net/minecraftforge/network/client/package-info.java
@@ -1,4 +1,0 @@
-@ApiStatus.Internal
-package net.minecraftforge.network.client;
-
-import org.jetbrains.annotations.ApiStatus;