+++ CHANGELOG.md
@@ -3,6 +3,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v5.0.0-1.19.3] - 2022-12-16
+This version includes major changes and refactors regarding the whole structure of Forge Config Api Port.
+### Added
+- Forge Config Api Port now includes a config file itself which includes the following options:
+  - An option to set the `defaultsconfigs` directory (just like Forge, not that it's too useful)
+  - An option to force server configs to generate in and load from the standard config directory (`.minecraft/config/`), so those configs are no longer world specific, but can be accessed much easier by users
+  - An option to manually prevent the custom command for opening configs from being registered, intended to be used when hosting a LAN world to allow clients without this mod to connect (changing the option requires a restart)
+### Changed
+- Classes not originally found in Forge now use a separate domain `fuzs.forgeconfigapiport` with a similar structure to Fabric Api (divided into `api`, `impl`, and `mixin`)
+- `fuzs.forgeconfigapiport` also is the new domain used for the Maven
+- The `api` package at `net.minecraftforge.api` has been moved to the new domain at `fuzs.forgeconfigapiport.api` and refactored:
+  - `net.minecraftforge.api.ModLoadingContext` -> `fuzs.forgeconfigapiport.api.config.v2.ForgeConfigRegistry`: Same as before, methods have been renamed from `registerConfig` to simply `register` though and registration needs an instance from `ForgeConfigRegistry#INSTANCE`.
+  - `net.minecraftforge.api.ConfigPaths` -> `fuzs.forgeconfigapiport.api.config.v2.ForgeConfigPaths`: Overhauled, includes helper methods for getting default paths for all config types, also provides the full file path, not just the directory name.
+  - `net.minecraftforge.api.fml.event.config.ModConfigEvents` -> `fuzs.forgeconfigapiport.api.config.v2.ModConfigEvents`: No changes, class has only been moved.
+- All implementation related classes have been compacted and moved to `fuzs.forgeconfigapiport.impl`
+- Mixin related classes have been moved to `fuzs.forgeconfigapiport.mixin`
+### Removed
+- This version also comes with many removals, mainly the WIP Forge config screens have been removed, as they were barely functional and the PR on Forge's GitHub has seemingly been abandoned. As an alternative for in-game configuration, Forge Config Api Port includes default support for the [Configured (Fabric)](https://www.curseforge.com/minecraft/mc-mods/configured-fabric) mod.
+
 ## [v4.2.9-1.19.2] - 2022-12-07
 ### Fixed
 - Fixed maven publication depending on Night Config as a mod at runtime